Chisel3 upgrade TileLink Xbar.scala, fold in Xbar_ACancel
